+++ src/utilities.jl
@@ -171,7 +171,9 @@ for AType in (:AbstractArray, :AbstractSciMLOperator)
     end
 end
 
+# alias of abstract types
 const FloatOrComplex = Union{AbstractFloat, Complex}
+const VectorOrTuple{T} = Union{AbstractVector{T}, NTuple{N, T} where {N}}
 
 # functions for getting Float or Complex element type
 _float_type(::AbstractArray{T}) where {T <: Number} = _float_type(T)