+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/qutip/QuantumToolbox.jl/tree/main)
 
-
+- Restructure `Dimensions` and `Space` to align with `qutip` (`Python`) ([#659]). Note that this update basically overwrites many changes made in PR [#600]:
+  - Replace `ProductDimensions` and `HilbertSpace` structures with `Dimensions` + `Space`/`TensorSpace`/`LiouvilleSpace`.
+  - Remove `get_hilbert_size` and `get_liouville_size`, and use `get_size` instead.
 
 ## [v0.44.0]
 Release date: 2026-03-11
@@ -456,6 +458,7 @@ Release date: 2024-11-13
 [#653]: https://github.com/qutip/QuantumToolbox.jl/issues/653
 [#656]: https://github.com/qutip/QuantumToolbox.jl/issues/656
 [#657]: https://github.com/qutip/QuantumToolbox.jl/issues/657
+[#659]: https://github.com/qutip/QuantumToolbox.jl/issues/659
 [#667]: https://github.com/qutip/QuantumToolbox.jl/issues/667
 [#669]: https://github.com/qutip/QuantumToolbox.jl/issues/669
 [#670]: https://github.com/qutip/QuantumToolbox.jl/issues/670